+++ tests/testthat/test-gmm.R
@@ -271,107 +271,53 @@ testthat::test_that("in case that the data includes NaN or Inf values, it return
 # GMM function
 #################
 
+test_GMMCluster <- function(res) {
+  if ('Error' %in% names(res)) {
+    expect_true( length(res) == 2)
+  } else {
+
+    expect_s3_class(res, "GMMCluster")
+    expect_equal(names(res),
+                 c("call", "centroids", "covariance_matrices",
+                   "weights", "Log_likelihood"))
+    expect_true(all(!sapply(res, is.null)))
+  }
+  
+}
 
 testthat::test_that("in case that the data is a matrix the result is a list of length 4 and the class is 'Gaussian Mixture Models' ", {
-  
-  res = GMM(X, 2, "maha_dist", "random_subset", 10, 10)
-  
-  if ('Error' %in% names(res)) {
-    
-    testthat::expect_true( length(res) == 2)}
-  
-  else {
-    
-    testthat::expect_true( length(res) == 4 && sum(names(res) %in% c("centroids", "covariance_matrices", "weights", "Log_likelihood")) == 4 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 4 && class(res) == "Gaussian Mixture Models" )
-  }
+  res <- GMM(X, 2, "maha_dist", "random_subset", 10, 10)
+  test_GMMCluster(res)
 })
 
-
 testthat::test_that("in case that the data is a data frame the result is a list of length 4 and the class is 'Gaussian Mixture Models' ", {
-  
-  res = GMM(dat, 2, "eucl_dist", "static_subset", 10, 10)
-  
-  if ('Error' %in% names(res)) {
-    
-    testthat::expect_true( length(res) == 2)}
-  
-  else {
-    
-    testthat::expect_true( length(res) == 4 && sum(names(res) %in% c("centroids", "covariance_matrices", "weights", "Log_likelihood")) == 4 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 4 && class(res) == "Gaussian Mixture Models" )
-  }
+  res <- GMM(dat, 2, "eucl_dist", "static_subset", 10, 10)
+  test_GMMCluster(res)
 })
 
 
 testthat::test_that("in case that the data is a matrix the result is a list of length 4 and the class is 'Gaussian Mixture Models' ", {
-  
-  res = GMM(X, 2, "maha_dist", "random_subset", 10, 10)
-  
-  if ('Error' %in% names(res)) {
-    
-    testthat::expect_true( length(res) == 2)}
-  
-  else {
-    
-    testthat::expect_true( length(res) == 4 && sum(names(res) %in% c("centroids", "covariance_matrices", "weights", "Log_likelihood")) == 4 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 4 && class(res) == "Gaussian Mixture Models" )
-  }
+  res <- GMM(X, 2, "maha_dist", "random_subset", 10, 10)
+  test_GMMCluster(res)
 })
 
 
 testthat::test_that("in case that the data is a data frame the result is a list of length 4 and the class is 'Gaussian Mixture Models' ", {
-  
-  res = GMM(dat, 2, "eucl_dist", "static_subset", 10, 10)
-  
-  if ('Error' %in% names(res)) {
-    
-    testthat::expect_true( length(res) == 2)}
-  
-  else {
-    
-    testthat::expect_true( length(res) == 4 && sum(names(res) %in% c("centroids", "covariance_matrices", "weights", "Log_likelihood")) == 4 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 4 && class(res) == "Gaussian Mixture Models" )
-  }
+  res <- GMM(dat, 2, "eucl_dist", "static_subset", 10, 10)
+  test_GMMCluster(res)
 })
 
 
 
 testthat::test_that("in case that the data is a matrix the result is a list of length 4 and the class is 'Gaussian Mixture Models' ", {
-  
-  res = GMM(X, 2, "maha_dist", "random_spread", 10, 10)
-  
-  if ('Error' %in% names(res)) {
-    
-    testthat::expect_true( length(res) == 2)}
-  
-  else {
-    
-    testthat::expect_true( length(res) == 4 && sum(names(res) %in% c("centroids", "covariance_matrices", "weights", "Log_likelihood")) == 4 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 4 && class(res) == "Gaussian Mixture Models" )
-  }
+  res <- GMM(X, 2, "maha_dist", "random_spread", 10, 10)
+  test_GMMCluster(res)
 })
 
 
 testthat::test_that("in case that the data is a data frame the result is a list of length 4 and the class is 'Gaussian Mixture Models' ", {
-  
-  res = GMM(dat, 2, "eucl_dist", "static_spread", 10, 10)
-  
-  if ('Error' %in% names(res)) {
-    
-    testthat::expect_true( length(res) == 2)}
-  
-  else {
-    
-    testthat::expect_true( length(res) == 4 && sum(names(res) %in% c("centroids", "covariance_matrices", "weights", "Log_likelihood")) == 4 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 4 && class(res) == "Gaussian Mixture Models" )
-  }
+  res <- GMM(dat, 2, "eucl_dist", "static_spread", 10, 10)
+  test_GMMCluster(res)
 })
 
 
@@ -446,14 +392,22 @@ testthat::test_that("in case that the determinant is zero the function returns a
 # predict_GMM function
 #######################
 
+testthat::test_that("GMM predict method works", {
+  gmm = GMM(dat, 3)
+  testthat::expect_equal(predict_GMM(dat,
+                                     CENTROIDS = gmm$centroids,
+                                     COVARIANCE = gmm$covariance_matrices,
+                                     WEIGHTS = gmm$weights)$cluster_labels,
+                         predict(gmm, dat))
+})
 
 testthat::test_that("in case that the data is a matrix the result is a list of length 3 and the class is 'Gaussian Mixture Models' ", {
   
   res = predict_GMM(X, CENTROIDS, COVARIANCE, WEIGHTS)
   
-  testthat::expect_true( length(res) == 3 && sum(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) == 3 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 3 && class(res) == "Gaussian Mixture Models")
+  testthat::expect_true(length(res) == 3 &&
+                        all(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) &&
+                        all(unlist(lapply(res, function(x) !is.null(x)))))
 })
 
 
@@ -461,9 +415,9 @@ testthat::test_that("in case that the data is a data frame the result is a list 
   
   res = predict_GMM(dat, CENTROIDS, COVARIANCE, WEIGHTS)
   
-  testthat::expect_true( length(res) == 3 && sum(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) == 3 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 3 && class(res) == "Gaussian Mixture Models" )
+  testthat::expect_true(length(res) == 3 &&
+                        all(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) &&
+                        all(unlist(lapply(res, function(x) !is.null(x)))))
 })
 
 
@@ -473,9 +427,9 @@ testthat::test_that("in case that the data is a matrix AND the CENTROIDS is a da
   
   res = predict_GMM(X, tmp_c, COVARIANCE, WEIGHTS)
   
-  testthat::expect_true( length(res) == 3 && sum(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) == 3 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 3 && class(res) == "Gaussian Mixture Models")
+  testthat::expect_true(length(res) == 3 &&
+                        all(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) &&
+                        all(unlist(lapply(res, function(x) !is.null(x)))))
 })
 
 
@@ -485,9 +439,9 @@ testthat::test_that("in case that the data is a matrix AND the COVARIANCE is a d
   
   res = predict_GMM(X, CENTROIDS, tmp_c, WEIGHTS)
   
-  testthat::expect_true( length(res) == 3 && sum(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) == 3 &&
-                           
-                           sum(unlist(lapply(res, function(x) !is.null(x)))) == 3 && class(res) == "Gaussian Mixture Models")
+  testthat::expect_true(length(res) == 3 &&
+                        all(names(res) %in% c("log_likelihood", "cluster_proba", "cluster_labels")) &&
+                        all(unlist(lapply(res, function(x) !is.null(x)))))
 })
 
 
@@ -632,7 +586,7 @@ testthat::test_that("in case that the data is a matrix the result is a vector an
   
   else {
     
-    testthat::expect_true( length(res) == Nr_clusters && class(res) == "Gaussian Mixture Models" )
+    testthat::expect_true( length(res) == Nr_clusters && inherits(res, "Gaussian Mixture Models") )
   }
 })
 
@@ -650,7 +604,7 @@ testthat::test_that("in case that the data is a data frame the result is a vecto
   
   else {
     
-    testthat::expect_true( length(res) == Nr_clusters && class(res) == "Gaussian Mixture Models" )
+    testthat::expect_true( length(res) == Nr_clusters && inherits(res, "Gaussian Mixture Models") )
   }
 })
 
@@ -668,7 +622,7 @@ testthat::test_that("in case of different parameters the result is a vector and 
   
   else {
     
-    testthat::expect_true( length(res) == Nr_clusters && class(res) == "Gaussian Mixture Models" )
+    testthat::expect_true( length(res) == Nr_clusters && inherits(res, "Gaussian Mixture Models") )
   }
 })
 
@@ -686,7 +640,7 @@ testthat::test_that("in case of a contiguous-vector it returns the correct outpu
   
   else {
     
-    testthat::expect_true( length(res) == length(Nr_clusters) && class(res) == "Gaussian Mixture Models" )
+    testthat::expect_true( length(res) == length(Nr_clusters) && inherits(res, "Gaussian Mixture Models") )
   }
 })
 
@@ -704,6 +658,6 @@ testthat::test_that("in case of a non-contiguous-vector it returns the correct o
   
   else {
     
-    testthat::expect_true( length(res) == length(Nr_clusters) && class(res) == "Gaussian Mixture Models" )
+    testthat::expect_true( length(res) == length(Nr_clusters) && inherits(res, "Gaussian Mixture Models") )
   }
 })