+++ tests/testthat/test-medoids.R
@@ -108,7 +108,7 @@ testthat::test_that("in case that the data is a matrix, it returns the correct o
   cm = Cluster_Medoids(X, clusters = 2, distance_metric = 'euclidean', swap_phase = TRUE, fuzzy = T)
 
   testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "best_dissimilarity", "dissimilarity_matrix",
-                                          "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && class(cm) == "cluster medoids silhouette" && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) &&
+                                          "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && inherits(cm, "cluster medoids silhouette") && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) &&
                            is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.data.frame(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.data.frame(cm$clustering_stats) )
 })
 
@@ -121,7 +121,7 @@ testthat::test_that("in case that the data is a dissimilarity matrix, it returns
   cm = Cluster_Medoids(dism_mat, clusters = 2, distance_metric = 'euclidean', swap_phase = TRUE, fuzzy = T)
 
   testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "best_dissimilarity", "dissimilarity_matrix",
-                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && class(cm) == "cluster medoids silhouette" && is.vector(cm$medoids) && is.vector(cm$medoid_indices) &&
+                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && inherits(cm, "cluster medoids silhouette") && is.vector(cm$medoids) && is.vector(cm$medoid_indices) &&
                            is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.data.frame(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.data.frame(cm$clustering_stats) )
 })
 
@@ -132,7 +132,7 @@ testthat::test_that("in case that the data is a matrix, it returns the correct o
   cm = Cluster_Medoids(dat, clusters = 2, distance_metric = 'euclidean', swap_phase = TRUE, fuzzy = T)
 
   testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "best_dissimilarity", "dissimilarity_matrix",
-                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && class(cm) == "cluster medoids silhouette" && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) &&
+                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && inherits(cm, "cluster medoids silhouette") && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) &&
                            is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.data.frame(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.data.frame(cm$clustering_stats) )
 })
 
@@ -142,7 +142,7 @@ testthat::test_that("in case that clusters = 1, it returns the correct output", 
   cm = Cluster_Medoids(dat, clusters = 1, distance_metric = 'euclidean', swap_phase = TRUE, fuzzy = T)
 
   testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "best_dissimilarity", "dissimilarity_matrix",
-                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && class(cm) == "cluster medoids silhouette" && is.vector(cm$medoids) && is.vector(cm$medoid_indices) &&
+                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats")) == 8 && inherits(cm, "cluster medoids silhouette") && is.vector(cm$medoids) && is.vector(cm$medoid_indices) &&
                            is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.null(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.null(cm$clustering_stats) )
 })
 
@@ -285,40 +285,40 @@ testthat::test_that("in case that the data includes NaN or Inf values, it return
 #########################
 
 
+test_MedoidsCluster <- function(cm, nclust) {
+  expect_true(all(c("medoids", "medoid_indices", "sample_indices", "best_dissimilarity",
+                    "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats",
+                    "dissimilarity_matrix") %in% names(cm)))
+  expect_is(cm, "MedoidsCluster")
+  expect_is(cm$medoids, "matrix")
+  expect_is(cm$medoid_indices, "numeric")
+  expect_is(cm$sample_indices, "numeric")
+  expect_is(cm$best_dissimilarity, "numeric")
+  expect_is(cm$clusters, "numeric")
+  expect_is(cm$fuzzy_probs, "matrix")
+  expect_is(cm$clustering_stats, "data.frame")
+  if (nclust > 1) {
+    expect_is(cm$silhouette_matrix, "data.frame")
+  } else {
+    expect_null(cm$silhouette_matrix)
+  }
+}
 
 testthat::test_that("in case that the data is a matrix, it returns the correct output", {
-
-  cm = Clara_Medoids(X, clusters = 2, samples = 5, sample_size = 0.2, swap_phase = TRUE, fuzzy = T)
-
-  testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "sample_indices", "best_dissimilarity",
-                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats",
-                                              "dissimilarity_matrix")) == 9 && class(cm) == "cluster medoids silhouette" && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) && is.vector(cm$sample_indices) &&
-                           is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.data.frame(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.data.frame(cm$clustering_stats) )
+  cm <- Clara_Medoids(X, clusters = 2, samples = 5, sample_size = 0.2, swap_phase = TRUE, fuzzy = T)
+  test_MedoidsCluster(cm, 2)
 })
 
 
 testthat::test_that("in case that the data is a data frame, it returns the correct output", {
-
-  cm = Clara_Medoids(dat, clusters = 2, samples = 5, sample_size = 0.2, swap_phase = TRUE, fuzzy = T)
-
-  testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "sample_indices", "best_dissimilarity",
-                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats",
-                                              "dissimilarity_matrix")) == 9 && class(cm) == "cluster medoids silhouette" && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) && is.vector(cm$sample_indices) &&
-                           is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.data.frame(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.data.frame(cm$clustering_stats) )
+  cm <- Clara_Medoids(dat, clusters = 2, samples = 5, sample_size = 0.2, swap_phase = TRUE, fuzzy = T)
+  test_MedoidsCluster(cm, 2)
 })
-
-
 
 testthat::test_that("in case that the clusters parameter is 1, it returns the correct output", {
-
-  cm = Clara_Medoids(dat, clusters = 1, samples = 5, sample_size = 0.2, swap_phase = TRUE, fuzzy = T)
-
-  testthat::expect_true( sum(names(cm) %in% c("medoids", "medoid_indices", "sample_indices", "best_dissimilarity",
-                                              "clusters", "silhouette_matrix", "fuzzy_probs", "clustering_stats",
-                                              "dissimilarity_matrix")) == 9 && class(cm) == "cluster medoids silhouette" && is.matrix(cm$medoids) && is.vector(cm$medoid_indices) && is.vector(cm$sample_indices) &&
-                           is.numeric(cm$best_dissimilarity) && is.vector(cm$clusters) && is.null(cm$silhouette_matrix) && is.matrix(cm$fuzzy_probs) && is.data.frame(cm$clustering_stats) )
+  cm <- Clara_Medoids(dat, clusters = 1, samples = 5, sample_size = 0.2, swap_phase = TRUE, fuzzy = T)
+  test_MedoidsCluster(cm, 1)
 })
-
 
 
 #################################
@@ -327,12 +327,9 @@ testthat::test_that("in case that the clusters parameter is 1, it returns the co
 
 
 testthat::test_that("in case that the data is not a matrix or a data frame, it returns an error", {
-
-  cm = Cluster_Medoids(X, clusters = 2, distance_metric = 'euclidean', swap_phase = TRUE, fuzzy = T)
-
-  tmp_x = list(X)
-
-  testthat::expect_error( predict_Medoids(tmp_x, MEDOIDS = cm$medoids, 'euclidean', fuzzy = TRUE) )
+  cm <- Cluster_Medoids(X, clusters = 2, distance_metric = 'euclidean', swap_phase = TRUE, fuzzy = T)
+  tmp_x <- list(X)
+  testthat::expect_error(predict_Medoids(tmp_x, MEDOIDS = cm$medoids, 'euclidean', fuzzy = TRUE) )
 })
 
 
@@ -408,7 +405,7 @@ testthat::test_that("in case that the data is a matrix, it returns the correct o
 
   pr = predict_Medoids(X, MEDOIDS = cm$medoids, 'euclidean', fuzzy = TRUE)
 
-  testthat::expect_true( sum(names(pr) %in% c("clusters", "fuzzy_clusters", "dissimilarity")) == 3 && class(pr) == "cluster medoids silhouette" && is.vector(pr$clusters) &&
+  testthat::expect_true( sum(names(pr) %in% c("clusters", "fuzzy_clusters", "dissimilarity")) == 3 && inherits(pr, "cluster medoids silhouette") && is.vector(pr$clusters) &&
                            is.matrix(pr$fuzzy_clusters) && is.numeric(pr$dissimilarity) )
 })
 
@@ -419,8 +416,10 @@ testthat::test_that("in case that the data is a data frame, it returns the corre
 
   pr = predict_Medoids(dat, MEDOIDS = cm$medoids, 'euclidean', fuzzy = TRUE)
 
-  testthat::expect_true( sum(names(pr) %in% c("clusters", "fuzzy_clusters", "dissimilarity")) == 3 && class(pr) == "cluster medoids silhouette" && is.vector(pr$clusters) &&
-                           is.matrix(pr$fuzzy_clusters) && is.numeric(pr$dissimilarity) )
+  testthat::expect_true(sum(names(pr) %in% c("clusters", "fuzzy_clusters", "dissimilarity")) == 3 &&
+                        inherits(pr, "cluster medoids silhouette") && is.vector(pr$clusters) &&
+                        is.matrix(pr$fuzzy_clusters) && is.numeric(pr$dissimilarity))
+  testthat::expect_equal(pr$clusters, predict(cm, dat))
 })
 
 
@@ -430,8 +429,9 @@ testthat::test_that("in case that MEDOIDS is a data frame, it returns the correc
 
   pr = predict_Medoids(X, MEDOIDS = tmp_cm, 'euclidean', fuzzy = TRUE)
 
-  testthat::expect_true( sum(names(pr) %in% c("clusters", "fuzzy_clusters", "dissimilarity")) == 3 && class(pr) == "cluster medoids silhouette" && is.vector(pr$clusters) &&
-                           is.matrix(pr$fuzzy_clusters) && is.numeric(pr$dissimilarity) )
+  testthat::expect_true(sum(names(pr) %in% c("clusters", "fuzzy_clusters", "dissimilarity")) == 3 &&
+                        inherits(pr, "cluster medoids silhouette") && is.vector(pr$clusters) &&
+                        is.matrix(pr$fuzzy_clusters) && is.numeric(pr$dissimilarity))
 })
 
 
@@ -604,7 +604,7 @@ testthat::test_that("in case that the data is a matrix, it returns the correct o
 
   opt_md = Optimal_Clusters_Medoids(X, max_clusters = 10, 'euclidean', 'dissimilarity', plot_clusters = F)
 
-  testthat::expect_true( class(opt_md) == "cluster medoids silhouette" && mean(unlist(lapply(opt_md, length))[-1]) == 6 )
+  testthat::expect_true( mean(unlist(lapply(opt_md, length))[-1]) == 6 )
 })
 
 
@@ -617,7 +617,7 @@ testthat::test_that("in case that the data is a matrix, it returns the correct o
   clust_four = opt_md[[4]]$sum_intra_dissim == opt_md1[[2]]$sum_intra_dissim
   clust_six = opt_md[[6]]$avg_width_silhouette == opt_md1[[3]]$avg_width_silhouette
 
-  testthat::expect_true( class(opt_md) == "cluster medoids silhouette" && mean(unlist(lapply(opt_md, length))[-1]) == 6 && all(clust_two, clust_four, clust_six) )
+  testthat::expect_true( mean(unlist(lapply(opt_md, length))[-1]) == 6 && all(clust_two, clust_four, clust_six) )
 })
 
 
@@ -625,7 +625,7 @@ testthat::test_that("in case that the data is a data frame, it returns the corre
 
   opt_md = Optimal_Clusters_Medoids(dat, max_clusters = 10, 'euclidean', 'dissimilarity', plot_clusters = FALSE)
 
-  testthat::expect_true( class(opt_md) == "cluster medoids silhouette" && mean(unlist(lapply(opt_md, length))[-1]) == 6 )
+  testthat::expect_true( mean(unlist(lapply(opt_md, length))[-1]) == 6 )
 })
 
 
@@ -639,26 +639,21 @@ testthat::test_that("in case of Cluster_Medoids for different parameter settings
 
     opt_md = Optimal_Clusters_Medoids(dat, max_clusters = 10, 'euclidean', tmp[i], plot_clusters = F)
 
-    res[i] = (inherits(opt_md, 'cluster medoids silhouette') && mean(unlist(lapply(opt_md, length))[-1]) == 6)
+    testthat::expect_equal(mean(unlist(lapply(opt_md, length))[-1]), 6)
   }
-
-  testthat::expect_true( sum(res) == length(tmp) )
+  
 })
 
 
 testthat::test_that("in case of Clara_Medoids for different parameter settings, it returns the correct output", {
 
-  tmp = c("silhouette","dissimilarity")
-
+  tmp = c("silhouette", "dissimilarity")
   res = rep(NA, length(tmp))
 
   for (i in 1:length(tmp)) {
-
     opt_md = Optimal_Clusters_Medoids(dat, max_clusters = 10, 'euclidean', tmp[i], clara_samples = 5, clara_sample_size = 0.2, plot_clusters = F)
-
-    res[i] = (inherits(opt_md, 'cluster medoids silhouette') && mean(unlist(lapply(opt_md, length))[-1]) == 6)
+    testthat::expect_equal(mean(unlist(lapply(opt_md, length))[-1]), 6)
   }
-
-  testthat::expect_true( sum(res) == length(tmp) )
+  
 })